Remove props on PollerMenu who can be used like a consts
