+++ www/front_src/src/Header/PollerMenu/models.ts
@@ -1,0 +1,9 @@
+export interface Issue {
+  critical: number;
+  total: number;
+  warning: number;
+}
+
+export interface Issues {
+  [key: string]: Issue;
+}
