+++ www/front_src/src/Header/index.tsx
@@ -13,14 +13,11 @@ import HostStatusCounter from './RessourceStatusCounter/Host';
 const HookComponent = Hook as unknown as (props) => JSX.Element;
 
 const Header = (): JSX.Element => {
-  const pollerListIssues =
-    'internal.php?object=centreon_topcounter&action=pollersListIssues';
-
   return (
     <header className={styles.header}>
       <div className={styles['header-icons']}>
         <div className={classnames(styles.wrap, styles['wrap-left'])}>
-          <PollerMenu endpoint={pollerListIssues} loaderWidth={27} />
+          <PollerMenu />
         </div>
         <div className={classnames(styles.wrap, styles['wrap-right'])}>
           <HookComponent path="/header/topCounter" />
