+++ www/front_src/src/Resources/Details/tabs/Details/CommandLine/index.tsx
@@ -2,29 +2,31 @@ import * as React from 'react';
 
 import { isNil } from 'ramda';
 
-import { makeStyles, Typography } from '@material-ui/core';
+import { createStyles, makeStyles, Typography } from '@material-ui/core';
 
 import { getCommandsWithArguments } from './utils';
 
-const useStyles = makeStyles((theme) => ({
-  argument: {
-    marginRight: theme.spacing(0.5),
-  },
-  argumentWithValue: {
-    display: 'flex',
-    marginLeft: theme.spacing(1),
-  },
-  command: {
-    fontWeight: theme.typography.fontWeightBold,
-  },
-  pipe: {
-    marginRight: theme.spacing(1),
-  },
-  pipedCommand: {
-    display: 'flex',
-    flexDirection: 'row',
-  },
-}));
+const useStyles = makeStyles((theme) =>
+  createStyles({
+    argument: {
+      marginRight: theme.spacing(0.5),
+    },
+    argumentWithValue: {
+      display: 'flex',
+      marginLeft: theme.spacing(1),
+    },
+    command: {
+      fontWeight: 'bold',
+    },
+    pipe: {
+      marginRight: theme.spacing(1),
+    },
+    pipedCommand: {
+      display: 'flex',
+      flexDirection: 'row',
+    },
+  }),
+);
 
 interface Props {
   commandLine: string;
